requirements.txt
astroid==2.3.3
click==7.1.2
colorama==0.4.3
fastapi==0.47.1
gunicorn==19.9.0
h11==0.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
six==1.14.0
starlette==0.12.9
uvicorn==0.11.2
google-cloud-firestore==1.9.0
firebase-admin==4.4.0
websockets==8.1
requests==2.23.0
wrapt==1.11.2
typing==3.7.4
pydantic==1.7.2
pycoingecko==1.4.0
pytz==2020.4
python-dateutil==2.8.1
tensorflow==2.3.0
scikit-learn==0.23.2